fix: remove es2018 and examples conditional importing
