+++ test/functional/examples/create_index.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const setupDatabase = require('../functional/shared').setupDatabase;
+const setupDatabase = require('../shared').setupDatabase;
 
 describe('examples.createIndex:', function () {
   let client;
