+++ test/functional/examples/project_fields_from_query_results.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const setupDatabase = require('../functional/shared').setupDatabase;
+const setupDatabase = require('../shared').setupDatabase;
 const expect = require('chai').expect;
 
 describe('examples(project-fields-from-query):', function () {
