+++ test/functional/examples/query_arrays.test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const setupDatabase = require('../functional/shared').setupDatabase;
+const setupDatabase = require('../shared').setupDatabase;
 const expect = require('chai').expect;
 
 describe('examples(query-arrays):', function () {
