+++ test/functional/examples/transactions.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const setupDatabase = require('../functional/shared').setupDatabase;
-const MongoClient = require('../../src').MongoClient;
+const setupDatabase = require('../shared').setupDatabase;
+const MongoClient = require('../../../src').MongoClient;
 
 describe('examples(transactions):', function () {
   let client;
